requirements.txt
Django>=5.0
gunicorn
python-dotenv
dj-database-url
psycopg2-binary
whitenoise